Remove TreeGrid PoC, re-merge parent components
For actual non-PoC grid implementation,
composition will be used.
