+++ vitest.config.js
@@ -8,7 +8,7 @@ export default defineConfig({
   test: {
     coverage: {
       include: ["src/**/*"],
-      exclude: ["src/stories/**/*", "src/components/TreeGrid*"],
+      exclude: ["src/stories/**/*"],
       all: true,
       statements: 90,
     },
